# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Worldwide Rapid Task For Money - Verification Automation Engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/activity-feed:
    get:
      operationId: getActivityFeed
      tags: [dashboard]
      summary: Recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
...
components:
  schemas:
...
    ActivityEvent:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [task_completed, task_rejected, campaign_created, campaign_paused, withdrawal_requested, user_registered, manual_review_needed, automation_milestone]
        description:
          type: string
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        campaignId:
          type: ["integer", "null"]
        taskId:
          type: ["integer", "null"]
        amount:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
...